HOCS-3949: reorganise config properties
